move encode-only code out of dec_patch_dictionary.h
