+++ lib/jxl/enc_detect_dots.h
@@ -15,6 +15,7 @@
 
 #include "lib/jxl/base/data_parallel.h"
 #include "lib/jxl/dec_patch_dictionary.h"
+#include "lib/jxl/enc_patch_dictionary.h"
 #include "lib/jxl/image.h"
 
 namespace jxl {
